jumns-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
mangum>=0.19.0
google-cloud-firestore>=2.19.0
google-cloud-storage>=2.18.0
google-auth>=2.29.0
google-adk>=1.0.0
google-genai>=1.0.0
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.10.0
numpy>=1.26.0
faiss-cpu>=1.8.0
websockets>=13.0.0
